Seems to make the UI 4:3 so it's aspect corrected
Seems to make the UI 4:3 from 1:1 to correct the aspect ratio. If 1 sets the height to 100% of the width, 0.75 could be 75% of the width. I was just messing with numbers trying to change something else, and I knew this was a requested fix. I have no idea what I'm doing.